separated styles for start screen, removed alpine
